+++ src/main/resources/META-INF/plexus/components.xml
@@ -11,7 +11,7 @@
           <lifecycle>
             <id>default</id>
             <phases>
-              <validate>org.jenkins-ci.tools:maven-hpi-plugin:validate</validate>
+              <validate>org.jenkins-ci.tools:maven-hpi-plugin:validate,org.jenkins-ci.tools:maven-hpi-plugin:validate-hpi</validate>
               <process-resources>org.apache.maven.plugins:maven-resources-plugin:resources</process-resources>
               <compile>org.apache.maven.plugins:maven-compiler-plugin:compile</compile>
               <process-classes>org.kohsuke:access-modifier-checker:enforce</process-classes>
